pyproject.toml
[project]
name = "barbershop"
version = "0.1.0"
description = ""
authors = [
    {name = "Vadim",email = "kozlov67b@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-telegram-bot (>=20.0)"
]

[tool.poetry]
packages = [{include = "barbershop", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
